fund-agent/requirements.txt
fastapi
uvicorn
langchain
python-dotenv
mysql-connector-python
pymysql
sqlalchemy
langchain-community
dashscope